requirements.txt
Flask==2.1.1
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.27.1
selenium==4.1.5
webdriver_manager==3.5.4
dnspython==2.2.1
gunicorn==20.1.0
